requirements.txt
fastapi
uvicorn[standard]
gunicorn
starlette
python-multipart
pillow
torch
torchvision
requests
python-dotenv
